+++ crates/rust-client/src/store/sqlite_store/store.sql
@@ -78,8 +78,8 @@ CREATE TABLE transactions (
     details BLOB NOT NULL,                           -- Serialized transaction details
     script_root TEXT,                                -- Transaction script root
     block_num UNSIGNED BIG INT,                      -- Block number for the block against which the transaction was executed.
-    commit_height UNSIGNED BIG INT NULL,             -- Block number of the block at which the transaction was included in the chain.
-    discard_cause BLOB NULL,                         -- Serialized cause of the discarded transaction
+    status_variant INT NOT NULL,                     -- Status variant identifier
+    status BLOB NOT NULL,                            -- Serialized transaction status
     FOREIGN KEY (script_root) REFERENCES transaction_scripts(script_root),
     PRIMARY KEY (id)
 );