+++ crates/rust-client/src/store/web_store/transaction/js_bindings.rs
@@ -25,8 +25,8 @@ extern "C" {
         transaction_id: String,
         details: Vec<u8>,
         block_num: String,
-        script_root: Option<Vec<u8>>,
-        committed: Option<String>,
-        discard_cause: Option<Vec<u8>>,
+        statusVariant: u8,
+        status: Vec<u8>,
+        scriptRoot: Option<Vec<u8>>,
     ) -> js_sys::Promise;
 }