+++ crates/rust-client/src/store/web_store/ts/sync.ts
@@ -132,9 +132,9 @@ interface SerializedTransactionData {
   id: string;
   details: Uint8Array;
   blockNum: string;
-  scriptRoot: Uint8Array;
-  commitHeight: string;
-  discardCause?: Uint8Array;
+  scriptRoot?: Uint8Array;
+  statusVariant: number;
+  status: Uint8Array;
   txScript?: Uint8Array;
 }
 
@@ -229,20 +229,27 @@ export async function applyStateSync(stateUpdate: JsStateSyncUpdate) {
   // Promises to insert each transaction update.
   let transactionWriteOp = Promise.all(
     transactionUpdates.map((transactionRecord) => {
-      return Promise.all([
-        insertTransactionScript(
-          transactionRecord.scriptRoot,
-          transactionRecord.txScript
-        ),
+      let promises = [
         upsertTransactionRecord(
           transactionRecord.id,
           transactionRecord.details,
           transactionRecord.blockNum,
-          transactionRecord.scriptRoot,
-          transactionRecord.commitHeight,
-          transactionRecord.discardCause
+          transactionRecord.statusVariant,
+          transactionRecord.status,
+          transactionRecord.scriptRoot
         ),
-      ]);
+      ];
+
+      if (transactionRecord.scriptRoot && transactionRecord.txScript) {
+        promises.push(
+          insertTransactionScript(
+            transactionRecord.scriptRoot,
+            transactionRecord.txScript
+          )
+        );
+      }
+
+      return Promise.all(promises);
     })
   );